pyproject.toml
[tool.poetry]
name = "rag-mkdocs"
version = "1.0.0"
description = "RAG assistant for MkDocs documentation"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.0"
langchain-openai = "^0.1.0"
langchain-community = "^0.2.0"
langchain-text-splitters = "^0.2.0"
faiss-cpu = "1.8.0"
fastapi = "0.110.0"
uvicorn = "0.29.0"
python-dotenv = "1.0.1"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.0"
httpx = "^0.27.0"
email-validator = "^2.1.0"
jinja2 = "^3.1.6"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
prometheus-client = "^0.20.0"
pytest = "^9.0.2"
pytest-asyncio = "^0.23.0"
ruff = "^0.8.0"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "asyncio: marks tests as async tests",
    "offline: marks tests as offline tests (no network/API keys required)",
    "online: marks tests as online tests (require running server)",
]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [
  "tests",
  ".pytest_cache",
  ".ruff_cache",
  ".mypy_cache",
  ".venv",
  "build",
  "dist",
  "site",
  ".artifacts",
  "var",
  "logs",
  "coverage",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # длину строки контролируем форматтером/прагматично

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

